requirements.txt
discord.py==2.3.2
Pillow==10.1.0
aiohttp==3.9.1
python-dotenv==1.0.0
